주소 목록 리스트 리펙토링
