+++ src/App.jsx
@@ -1,18 +1,6 @@
-export default function App() {
-  const addresses = [
-    {
-      id: 'eaebe552-ae57-4ac8-8c3c-1cb3d9dd9e54',
-      name: '서울',
-    },
-    {
-      id: '2ae6620b-e902-4fc4-ac6d-f4a733acf2db',
-      name: '대전',
-    },
-    {
-      id: 'dbe6172f-efb9-4e00-8cbc-21cbb5f2eaab',
-      name: '대구',
-    }];
+import AddressListContainer from './AddressListContainer';
 
+export default function App() {
   const categorires = [
     {
       id: '42281be6-d068-4654-8337-a24ff4cc8200',
@@ -37,10 +25,7 @@ export default function App() {
 
   return (
     <>
-      <ul>
-        {addresses.map(({ id, name }) => (<li key={id}>{name}</li>))}
-      </ul>
-
+      <AddressListContainer />
       <ul>
         {categorires.map(({ id, name }) => (<li key={id}>{name}</li>))}
       </ul>
